# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SparqMake API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creatives:
    get:
      operationId: getCreatives
      tags: [creatives]
      summary: Get all creatives
      parameters:
        - name: brandId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of creatives
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeListResponse"
    post:
      operationId: createCreative
      tags: [creatives]
      summary: Create a creative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreativeInput"
      responses:
        "201":
          description: Creative created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creative"
  /creatives/{id}:
    get:
      operationId: getCreative
      tags: [creatives]
      summary: Get a creative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Creative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creative"
    put:
      operationId: updateCreative
      tags: [creatives]
      summary: Update a creative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreativeInput"
      responses:
        "200":
          description: Creative updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creative"
...
components:
  schemas:
...
    CreativeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Creative"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - data
...
    Creative:
      type: object
      properties:
        id:
          type: string
        brandId:
          type: string
        templateId:
          type: string
          nullable: true
        name:
          type: string
        status:
          type: string
        briefText:
          type: string
          nullable: true
        referenceUrl:
          type: string
          nullable: true
        selectedAssets:
          type: array
          items:
            type: object
            additionalProperties: true
        selectedHashtagSets:
          type: array
          items:
            type: string
          nullable: true
        sourceCreativeId:
          type: string
          nullable: true
        intent:
          type: string
          nullable: true
        createdBy:
          type: string
        reviewedBy:
          type: string
          nullable: true
        reviewComment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - brandId
        - name
        - status
        - selectedAssets
        - createdBy
        - createdAt
        - updatedAt
    CreateCreativeInput:
      type: object
      properties:
        brandId:
          type: string
        templateId:
          type: string
          nullable: true
        name:
          type: string
        briefText:
          type: string
          nullable: true
        referenceUrl:
          type: string
          nullable: true
        selectedAssets:
          type: array
          items:
            type: object
            additionalProperties: true
        selectedHashtagSets:
          type: array
          items:
            type: string
          nullable: true
        intent:
          type: string
          nullable: true
        createdBy:
          type: string
      required:
        - brandId
        - name
        - selectedAssets
        - createdBy
    UpdateCreativeInput:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        briefText:
          type: string
          nullable: true
        referenceUrl:
          type: string
          nullable: true
        templateId:
          type: string
          nullable: true
        selectedAssets:
          type: array
          items:
            type: object
            additionalProperties: true
        selectedHashtagSets:
          type: array
          items:
            type: string
          nullable: true
        intent:
          type: string
          nullable: true
        reviewedBy:
          type: string
          nullable: true
        reviewComment:
          type: string
          nullable: true
...