# lib/api-zod/src/generated/types/creative.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SparqMake API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreativeSelectedAssetsItem } from "./creativeSelectedAssetsItem";

export interface Creative {
  id: string;
  brandId: string;
  templateId?: string | null;
  name: string;
  status: string;
  briefText?: string | null;
  referenceUrl?: string | null;
  selectedAssets: CreativeSelectedAssetsItem[];
  selectedHashtagSets?: string[] | null;
  sourceCreativeId?: string | null;
  intent?: string | null;
  createdBy: string;
  reviewedBy?: string | null;
  reviewComment?: string | null;
  createdAt: Date;
  updatedAt: Date;
}