# lib/api-zod/src/generated/types/updateCreativeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SparqMake API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCreativeInputSelectedAssetsItem } from "./updateCreativeInputSelectedAssetsItem";

export interface UpdateCreativeInput {
  name?: string;
  status?: string;
  briefText?: string | null;
  referenceUrl?: string | null;
  templateId?: string | null;
  selectedAssets?: UpdateCreativeInputSelectedAssetsItem[];
  selectedHashtagSets?: string[] | null;
  intent?: string | null;
  reviewedBy?: string | null;
  reviewComment?: string | null;
}